pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.coverage.paths]
source = ["src", "*/site-packages"]
tests = ["tests", "*/tests"]

[tool.coverage.report]
fail_under = 0  # Prevent Tests workflow to fails even if coverage is below 100.
show_missing = true

[tool.coverage.run]
branch = true
source = ["super_resolution", "tests"]

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["PIL", "yaml", "torchvision.*", "skimage.*", "torch.*"]
ignore_missing_imports = true

[tool.poetry]
authors = ["Mathias Gonin <pubnul0@gmail.com>"]
classifiers = [
    "Development Status :: 1 - Planning",
]
description = "Super resolution project."
documentation = "https://super_resolution.readthedocs.io"
homepage = "https://github.com/sing-lab/super_resolution"
license = "MIT"
name = "super_resolution"
readme = "README.md"
repository = "https://github.com/sing-lab/super_resolution"
version = "1.0.0"

[tool.poetry.dependencies]
click = "^8.1.3"
Pillow = "^9.2.0"
python = ">=3.8,<4.0.0"
scikit-image = "^0.19.3"
tensorboard = "^2.9.1"
PyYAML = "^6.0"
torch = "^1.13.1"
torchvision = "^0.14.1"

[tool.poetry.dev-dependencies]
bandit = "1.7.2"  # Pinned version to avoid bugs.
black = ">=21.10b0"
codecov = "^2.1.12"
coverage = {extras = ["toml"], version = "^6.5.0"}
darglint = "^1.8.1"
flake8 = "^3.7.9"
flake8-bandit = "^2.1.2"
flake8-black = "^0.3.3"
flake8-bugbear = "^21.9.2"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
flake8-rst-docstrings = "^0.2.3"
furo = "^2022.6.4"
mypy = "^0.910"
nox = "^2022.8.7"
nox-poetry = "^1.0.2"
pep8-naming = "^0.12.1"
pre-commit = "^2.15.0"
pre-commit-hooks = "^4.0.1"
Pygments = "^2.10.0"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pyupgrade = "^2.29.1"
safety = "^2.3.5"
Sphinx = "^5.2.3"
sphinx-autobuild = ">=2021.3.14"
sphinx-autodoc-typehints = "^1.19.4"
sphinx-click = "^3.0.2"
streamlit = "^1.11.0"
streamlit-image-comparison = "^0.0.2"
toml-sort = "^0.20.1"
typeguard = "^2.13.2"
xdoctest = {extras = ["colors"], version = "^0.15.10"}

[tool.poetry.scripts]
super_resolution = "super_resolution.main:main"  # Run project via poetry run super_resolution {config_path}

[tool.poetry.urls]
Changelog = "https://github.com/sing-lab/super_resolution/releases"

[tool.poetry.group.dev.dependencies]
codecov = "^2.1.12"

[tool.tomlsort]
all = true
ignore_case = true
in_place = true

[virtualenvs]
in-project = true
